demo_chatbot/requirements.txt
langchain==0.1.3
langchain-community==0.0.16
faiss-cpu==1.7.4
sentence-transformers==2.2.2
ollama==0.1.6
numpy==1.24.3
beautifulsoup4==4.12.2
argparse==1.4.0
gradio==4.13.0
